test/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
google-cloud-bigquery==3.11.4
google-cloud-pubsub==2.18.4
google-cloud-logging==3.5.0
google-cloud-storage==2.10.0
